refactor: change code structure according to the architecture
