+++ project-e-dukate-frontend/src/components/Appointments/index.ts
@@ -1,0 +1,2 @@
+export { AppointmentTable } from './AppointmentTable';
+export { AddAppointmentDialog } from './AddAppointmentDialog';
