+++ project-e-dukate-frontend/src/pages/Appointments/index.ts
@@ -1,0 +1,1 @@
+export { default as Appointments } from './Appointments';
